streamlit_app/requirements.txt
streamlit==1.40.1
opencv-python==4.12.0.88
deepface==0.0.93
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.5
pillow==10.4.0
requests==2.32.4
fpdf==2.5.7
